requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
requests>=2.25.0

# Data visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# PDF generation
reportlab>=3.6.0

# Financial data
yfinance>=0.1.70

# Development and testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.6.0
flake8>=3.9.0

# Environment management
python-dotenv>=0.19.0

# Database migrations (optional)
alembic>=1.7.0

# Logging
structlog>=21.1.0
